requirements.txt
oracledb
pandas
customtkinter
tkinterdnd2
pyzipper
cryptography
beautifulsoup4
requests
Pillow
pyinstaller
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyautogui
python-dotenv